Move code into actual files
The goal is to have a toml file for configuration and to build a command
line tool.

The program runs and writes valid html, but doesn't use the toml file
right now.  It also can't run as a command line tool yet.

Need to also fix the logic so that a single operator has at most two
shifts in a row.