Improve relative weights choice of operator
In a simple world, each operator would be equally likely of being picked
for a shift.  However, at least two things complicate this approach.
Some operators are unavailable for some shifts, like choir in the
morning.  This immediately makes the distribution non-uniform.  But we
would still like to distribute shifts equally among all operators over a
given time period.

The first solution was to assume uniform distribution.  To correct for
skewing, the program would repeatedly generate a schedule until the
difference between the operator with the most shifts and the operator
with least shifts was two or less.  This kind of worked.  Even though
the shifts distributed fairly equally overall, it was too common to have
operators to have streaks of 2 or more shifts in a row.

The second solution was to introduce relative weights.  The basic
algorithm was taken from the Python docs under random.choice() and, in
Python 3.6, it is called random.choices().  Basically, you give the
function a list of things to choose from, but you also provide a
relative weight for each item in the list:  Items with a higher weight
are more likely to be picked.  In this case, the items are the names of
the operators.

I tweaked the algorithm so that the weight of the operator who was last
picked decreases by some factor relative to the other operators.  This
makes it increasingly unlikely to produce streaks.  I did this by
multiplying the weights of all the other operators by some factor
greater than 1 (Decreasing the weight of the last operator selected is
more straightforward, but makes the math a little more complicated.  You
also can't achieve as wide a range of weights and would introduce
floating point numbers).  I also scaled down the weights by dividing
each weight by the smallest weight.  This keeps the weights from growing
too large.

After some initial testing by multiplying by a factor of 10, it seems
the current algorithm distributes shifts pretty equally overall, yet
keeps streaks to a length of two.