Add pyinstaller temporary directory
After building an executable, the path to the temporary directory where
stuff is extracted needs to be added to the html template files.  This
should only add the temp dir path if bundled.  If you just run the
program as a python script, it should still work like normal.

I kept getting a "No file found" errno 2 error because the executable
was looking in the directory where the executable was located.  Reading
the pyinstaller documentation sections "What PyInstaller Does and How It
Does It" and "Run-time Information" helped immensely.