Add build file for pyinstaller
This should bundle all code and data files into a single executable
file.  The extra flag tells pyinstaller to include the html template
files.

The executable should work whether or not the html-template directory is
in the working directory.  If the program is run as a normal Python
script, the html-template directory should be in the same directory as
the executable.