Create working build
Originally, I wanted to bundle the html template with the executable so
that the only thing to deploy would be the executable.  However, for
maximum flexibility, I decided not to.  Maybe in the future though.

I also changed the spec file location, so that it's easier to clean up
manually after a build.