Fix scaling problem
The largest relative weight was in the 1000000s.  The culprit was that
operators with no shifts were being added to the relative weights.

An operator with no shifts (eg Sun AM or Wed) may be added in order to
list contact information, but is not regularly scheduled.  Since that
particular operator was never picked, its weight kept increasing.

The fix only adds operators that have regular shifts.  The part for
listing contact information is unaffected.