Add date exceptions for operators
The immediate use case is for operators who are deacons.  They are
unavailable for the first Sunday PM shift of the month, but are
otherwise available.

Exceptions are enumerated as a list of lists where each inner list is
[date_string, shift_string].  An exception for the whole day can also be
specified by [date_string].

Adding a date range (for vacations) will need to be added.